third_party/WebKit/LayoutTests/http/tests/security/suborigins/crossorigin/suborigin-cross-origin-script-window-onerror-no-cors-expected.txt

CONSOLE ERROR: line 1: Uncaught SomeError
The test passes if 'window.onerror' gets sanitized information about an exception thrown in a script loaded from a suborigin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /SomeError/.test(msg) is false
PASS url is ""
PASS line is 0
PASS column is 0
PASS window.errorObject is null
PASS successfullyParsed is true

TEST COMPLETE
